Rename gateway function per PR feedback
