+++ packages/api/src/services/gateways/interfaces/gateway.ts
@@ -2,7 +2,7 @@ import type { User } from "@moj-bichard7/common/types/User"
 
 interface Gateway {
   fetchUserByUsername: (username: string) => Promise<User>
-  caseCanBeResubmitted: (username: string, caseId: number, forceIds: number[]) => Promise<boolean>
+  canCaseBeResubmitted: (username: string, caseId: number, forceIds: number[]) => Promise<boolean>
 }
 
 export default Gateway
