+++ packages/api/src/services/gateways/postgres/canCaseBeResubmitted.ts
@@ -1,13 +1,13 @@
 import type postgres from "postgres"
 
-export interface CaseCanBeResubmittedResult {
+export interface CanCaseBeResubmittedResult {
   locked_by_user: boolean
   case_in_force: boolean
   case_is_unresolved: boolean
 }
 
 export default async (sql: postgres.Sql, username: string, caseId: number, forceIds: number[]): Promise<boolean> => {
-  const [result]: [CaseCanBeResubmittedResult?] = await sql`
+  const [result]: [CanCaseBeResubmittedResult?] = await sql`
       SELECT
         (el.error_locked_by_id = ${username})::INTEGER as locked_by_user,
         (br7own.force_code(el.org_for_police_filter) = ANY((${forceIds}::SMALLINT[])))::INTEGER as case_in_force,
