+++ packages/api/src/services/gateways/postgresGateway.ts
@@ -1,7 +1,7 @@
 import type { User } from "@moj-bichard7/common/types/User"
 import postgresFactory from "../db/postgresFactory"
 import type Gateway from "./interfaces/gateway"
-import caseCanBeResubmitted from "./postgres/caseCanBeResubmitted"
+import caseCanBeResubmitted from "./postgres/canCaseBeResubmitted"
 import fetchUserByUsername from "./postgres/fetchUserByUsername"
 
 class PostgresGateway implements Gateway {
@@ -11,7 +11,7 @@ class PostgresGateway implements Gateway {
     return await fetchUserByUsername(this.db, username)
   }
 
-  async caseCanBeResubmitted(username: string, caseId: number, forceIds: number[]): Promise<boolean> {
+  async canCaseBeResubmitted(username: string, caseId: number, forceIds: number[]): Promise<boolean> {
     return await caseCanBeResubmitted(this.db, username, caseId, forceIds)
   }
 }
